moved NamedTupleVariable back to lists
