+++ torch/_dynamo/variables/__init__.py
@@ -97,6 +97,7 @@ from .lists import (
     BaseListVariable,
     ListIteratorVariable,
     ListVariable,
+    NamedTupleVariable,
     RangeVariable,
     SliceVariable,
     TupleIteratorVariable,
@@ -147,7 +148,6 @@ from .torch import TorchCtxManagerClassVariable, TorchInGraphFunctionVariable
 from .user_defined import (
     FrozenDataClassVariable,
     MutableMappingVariable,
-    NamedTupleVariable,
     RemovableHandleVariable,
     UserDefinedClassVariable,
     UserDefinedDictVariable,
