+++ torch/_dynamo/variables/builder.py
@@ -229,6 +229,7 @@ from .lists import (
     BaseListVariable,
     ListIteratorVariable,
     ListVariable,
+    NamedTupleVariable,
     RangeVariable,
     SizeVariable,
     SliceVariable,
@@ -293,7 +294,6 @@ from .user_defined import (
     IntWrapperVariable,
     KeyedJaggedTensorVariable,
     MutableMappingVariable,
-    NamedTupleVariable,
     SourcelessGraphModuleVariable,
     UserDefinedClassVariable,
     UserDefinedDictVariable,
